requirements.txt
pytest
pytest-cov
pre-commit
ruff
fastapi
pydantic
python-dotenv
PyJWT
passlib
python-multipart
httpx
supabase
email-validator
uvicorn
bcrypt
